+++ .github/workflows/test-phpunit.yml
@@ -41,7 +41,7 @@ jobs:
       fail-fast: false
       matrix:
         php-versions: ['7.2', '7.3', '7.4']
-        db-platforms: ['MySQLi', 'Postgre', 'SQLite3', 'Sqlsrv']
+        db-platforms: ['MySQLi', 'Postgre', 'SQLite3', 'SQLSRV']
 
     services:
       mysql:
@@ -77,7 +77,7 @@ jobs:
 
     steps:
       - name: Create database for MSSQL Server
-        if: ${{ matrix.db-platforms == 'Sqlsrv'}}
+        if: ${{ matrix.db-platforms == 'SQLSRV'}}
         run: sqlcmd -S localhost -U sa -P 1Secure*Password1 -Q "CREATE DATABASE test"
 
       - name: Checkout
